# Conflicts:
#	src/org/usfirst/frc/falcons6443/robot/Robot.java
#	src/org/usfirst/frc/falcons6443/robot/commands/TeleopMode.java
#	src/org/usfirst/frc/falcons6443/robot/subsystems/DriveTrainSystem.java
#	sysProps.xml

+++ src/org/usfirst/frc/falcons6443/robot/hardware/DriveEncoders.java
@@ -15,11 +15,11 @@ public class DriveEncoders {
         right.setReverseDirection(true);
     }
 
-    public double getLeftDistance(){
+    public int getLeftDistance(){
         return left.getRaw();
     }
 
-    public double getRightDistance(){
+    public int getRightDistance(){
         return right.getRaw();
     }
